REFACTOR: Move pushout complements of C-sets to `CSets` module.
This is where the other limit and colimit code is.
